Cleanup withColorSignal filters, simplify and reuse functions to add color signal classnames
