+++ packages/block-editor/src/filters/with-color-signal/index.js
@@ -1,312 +1,17 @@
-import _ from 'lodash';
-import classnames from 'classnames';
-
-import { Fragment } from '@wordpress/element';
-import { createHigherOrderComponent } from '@wordpress/compose';
-import { dispatch, select, subscribe } from '@wordpress/data';
-
 import { addFilter } from '@wordpress/hooks';
 
-import {
-  getParentVariation,
-  getSupports,
-} from "../../utils";
+import "./update-blocks";
 
-import {
-  computeColorSignal,
-  getAbsoluteColorVariation,
-  removeSiteVariationOffset,
-} from "@novablocks/utils";
+import withColorSignalAttributes from './with-color-signal-attributes';
+import withColorSignalsDeprecated from './with-color-signal-deprecated';
+import withColorSignalControls from './with-color-signal-controls';
+import withColorSignalEditClassnames from './with-color-signal-edit-classnames';
+import withColorSignalSaveClassnames from "./with-color-signal-save-classnames";
+import withColorSignalSaveDataAttributes from "./with-color-signal-save-data-attributes";
 
-import { useSupports } from "../../index";
-import InspectorControls from './inspector-controls';
-import attributes from './attributes.json';
-import altAttributes from './attributes-alt.json';
-
-const withColorSignalAttributes = ( settings, name ) => {
-
-  if ( ! settings?.supports?.novaBlocks?.colorSignal ) {
-    return settings;
-  }
-
-  return {
-    ...settings,
-    attributes: {
-      ...settings.attributes,
-      ...attributes,
-      ...( !! settings?.supports?.novaBlocks?.colorSignal?.altAttributes ? altAttributes : {} )
-    }
-  };
-}
-addFilter( 'blocks.registerBlockType', 'nova-blocks/with-color-signal-attributes', withColorSignalAttributes );
-
-const withColorSignalsDeprecation = ( settings, name ) => {
-
-  if ( ! settings?.supports?.novaBlocks?.colorSignal?.addOverlayColorDeprecatedMethod ) {
-    return settings;
-  }
-
-  return Object.assign( {}, settings, {
-    attributes: _.omit( settings.attributes, [ 'contentColor', 'overlayFilterStyle' ] ),
-    deprecated: [
-      {
-        attributes: settings.attributes,
-        isEligible( attributes ) {
-          return "undefined" === typeof attributes.paletteVariation;
-        },
-        migrate( attributes ) {
-          const { contentColor, overlayFilterStyle } = attributes;
-          const hadDarkText = !! contentColor && contentColor.search( '000' ) > -1;
-          const hadLightBackground = !! overlayFilterStyle && overlayFilterStyle === 'light';
-          const paletteVariation = hadDarkText || hadLightBackground ? '0' : '12';
-          const colorSignal = hadDarkText || hadLightBackground ? 0 : 3;
-
-          return {
-            ...attributes,
-            defaultsGenerated: true,
-            paletteVariation,
-            colorSignal,
-          };
-        },
-        save: settings.save,
-      }
-    ].concat( settings.deprecated ),
-  } );
-}
-addFilter( 'blocks.registerBlockType', 'nova-blocks/with-color-signal-deprecation', withColorSignalsDeprecation );
-
-const withColorSignalControls = createHigherOrderComponent( OriginalComponent => {
-
-  return props => {
-
-    const supports = useSupports( props.name );
-
-    if ( ! supports?.novaBlocks?.colorSignal ) {
-      return <OriginalComponent { ...props } />
-    }
-
-    return (
-      <Fragment>
-        <InspectorControls { ...props } />
-        <OriginalComponent { ...props } />
-      </Fragment>
-    );
-  }
-} );
+addFilter( 'blocks.registerBlockType', 'novablocks/with-color-signal-attributes', withColorSignalAttributes );
+addFilter( 'blocks.registerBlockType', 'novablocks/with-color-signal-deprecated', withColorSignalsDeprecated );
 addFilter( 'editor.BlockEdit', 'novablocks/with-color-signal-controls', withColorSignalControls );
-
-// Add Palette Classes in Edit
-const withPaletteClassname = createHigherOrderComponent( ( BlockListBlock ) => {
-
-  return ( props ) => {
-
-    const supports = useSupports( props.name );
-
-    if ( ! supports?.novaBlocks?.colorSignal?.paletteClassname ) {
-      return <BlockListBlock { ...props } />
-    }
-
-    const blockProps = {
-      ...props,
-      className: `${ props.className } sm-palette-${ props.attributes.palette } ${ props.attributes.useSourceColorAsReference ? 'sm-palette--shifted' : '' }`
-    } ;
-
-    return (
-      <BlockListBlock { ...blockProps } />
-    )
-  };
-}, "withPaletteClassname" );
-addFilter( 'editor.BlockListBlock', 'novablocks/with-palette-classname', withPaletteClassname );
-
-// Add Palette Classes on Save
-const applyPaletteFrontEndClasses = (extraProps, blockType, attributes) => {
-
-  const { palette, useSourceColorAsReference } = attributes;
-
-  const supports = getSupports( blockType.name );
-
-  if ( supports?.novaBlocks?.colorSignal?.paletteClassname ) {
-    extraProps.className = classnames(extraProps.className, `sm-palette-${ palette } ${ useSourceColorAsReference ? 'sm-palette--shifted' : '' }`)
-  }
-
-  return extraProps;
-}
-addFilter('blocks.getSaveContent.extraProps', 'novablocks-with-palette-classname-frontend', applyPaletteFrontEndClasses, 1);
-
-// Add Variation Classes on Edit
-const withVariationClassname = createHigherOrderComponent( ( BlockListBlock ) => {
-
-  return ( props ) => {
-
-    const supports = useSupports( props.name );
-
-    if ( ! supports?.novaBlocks?.colorSignal?.variationClassname ) {
-      return <BlockListBlock { ...props } />
-    }
-
-    const {
-      attributes: {
-        paletteVariation,
-        colorSignal,
-      },
-    } = props;
-
-    if ( colorSignal === 0 ) {
-      return <BlockListBlock { ...props } />
-    }
-
-    const className = classnames(
-      props.className,
-      `sm-variation-${ paletteVariation }`
-    );
-
-    const blockProps = { ...props, className };
-
-    return (
-      <BlockListBlock { ...blockProps } />
-    )
-  }
-}, "withVariationClassname" );
-
-const updateInnerBlocks = ( block ) => {
-
-  if ( Array.isArray( block.innerBlocks ) ) {
-    block.innerBlocks.forEach( innerBlock => {
-      updateBlock( innerBlock );
-    } )
-  }
-}
-
-const updateBlock = ( block ) => {
-
-  const supports = getSupports( block.name );
-
-  if ( supports?.novaBlocks?.colorSignal ) {
-    const { updateBlockAttributes } = dispatch( 'core/block-editor' );
-    const { attributes, clientId } = block;
-    const { colorSignal, paletteVariation, useSourceColorAsReference } = attributes;
-    const parentVariation = getParentVariation( clientId );
-    const absoluteVariation = getAbsoluteColorVariation( attributes );
-    const nextVariation = computeColorSignal( parentVariation, colorSignal, absoluteVariation );
-    const finalVariation = useSourceColorAsReference ? 1 : removeSiteVariationOffset( nextVariation );
-
-    if ( paletteVariation !== finalVariation ) {
-      updateBlockAttributes( clientId, {
-        paletteVariation: finalVariation,
-      } );
-    }
-  }
-
-  updateInnerBlocks( block );
-}
-
-addFilter( 'editor.BlockListBlock', 'novablocks/with-variation-classname', withVariationClassname );
-
-// Add Variation Classes on Save
-const applyVariationFrontEndClasses = (extraProps, blockType, attributes) => {
-
-  const { colorSignal, paletteVariation } = attributes;
-
-  const supports = getSupports( blockType.name );
-
-  if ( ! supports?.novaBlocks?.colorSignal?.variationClassname  ) {
-      return extraProps;
-  }
-
-  if ( colorSignal === 0 ) {
-    return extraProps;
-  }
-
-  extraProps.className = classnames( extraProps.className, `sm-variation-${ paletteVariation }` )
-
-  return extraProps;
-}
-addFilter('blocks.getSaveContent.extraProps', 'novablocks-with-variation-classname-frontend', applyVariationFrontEndClasses, 1);
-
-// Add Color Signal Classes
-const withColorSignalClassname = createHigherOrderComponent( ( BlockListBlock ) => {
-
-  return ( props ) => {
-
-    const supports = useSupports( props.name );
-
-    if ( ! supports?.novaBlocks?.colorSignal?.colorSignalClassname ) {
-      return <BlockListBlock { ...props } />
-    }
-
-    const blockProps = {
-      ...props,
-      className: `${ props.className } sm-color-signal-${ props.attributes.colorSignal }`
-    };
-
-    return (
-      <BlockListBlock { ...blockProps } />
-    )
-  }
-}, "withColorSignalClassname" );
-
-addFilter( 'editor.BlockListBlock', 'novablocks/with-color-signal-classname', withColorSignalClassname );
-
-// Add Color Signal Classes on Save
-const applyColorSignalFrontEndClasses = (extraProps, blockType, attributes) => {
-
-  const { colorSignal } = attributes;
-
-  const supports = getSupports( blockType.name );
-
-  if ( supports?.novaBlocks?.colorSignal?.colorSignalClassname ) {
-    extraProps.className = classnames( extraProps.className, `sm-color-signal-${ colorSignal }` )
-  }
-
-  return extraProps;
-}
-addFilter('blocks.getSaveContent.extraProps', 'novablocks-with-color-signal-classname-frontend', applyColorSignalFrontEndClasses, 1);
-
-// Add Color Signal Classes on Save
-const addColorSignalDataToSaveElement = ( element, blockType, attributes ) => {
-
-  const supports = getSupports( blockType.name );
-
-  if ( ! element || ! supports?.novaBlocks?.colorSignal ) {
-    return element;
-  }
-
-  const newProps = {
-    'data-palette': attributes?.palette,
-    'data-palette-variation': attributes?.paletteVariation,
-    'data-color-signal': attributes?.colorSignal,
-  }
-
-  if ( attributes.useSourceColorAsReference ) {
-    Object.assign( newProps, {
-      'data-use-source-color-as-reference': attributes?.useSourceColorAsReference,
-    } );
-  }
-
-  return Object.assign( {}, element, {
-    props: {
-      ...element?.props,
-      ...newProps,
-    }
-  } );
-}
-addFilter('blocks.getSaveElement', 'novablocks/add-color-signal-data-to-save-element', addColorSignalDataToSaveElement, 1);
-
-
-//
-const getBlockList = () => select( 'core/editor' ).getBlocks();
-
-let blockList = getBlockList();
-
-subscribe( () => {
-  const newBlockList = getBlockList();
-  const blockListChanged = newBlockList !== blockList;
-  blockList = newBlockList;
-
-  if ( blockListChanged ) {
-    // You can trigger here any behavior when the block list in the post changes.
-    blockList.forEach( ( block ) => {
-      updateBlock( block );
-    } );
-  }
-} );
+addFilter( 'editor.BlockListBlock', 'novablocks/with-color-signal-edit-classnames', withColorSignalEditClassnames );
+addFilter( 'blocks.getSaveContent.extraProps', 'novablocks/with-color-signal-save-classnames', withColorSignalSaveClassnames, 1);
+addFilter( 'blocks.getSaveElement', 'novablocks/add-color-signal-data-to-save-element', withColorSignalSaveDataAttributes, 1);
