+++ packages/block-editor/src/filters/with-color-signal/with-color-signal-attributes.js
@@ -1,0 +1,20 @@
+import attributes from "./attributes.json";
+import altAttributes from "./attributes-alt.json";
+
+const withColorSignalAttributes = ( settings, name ) => {
+
+  if ( ! settings?.supports?.novaBlocks?.colorSignal ) {
+    return settings;
+  }
+
+  return {
+    ...settings,
+    attributes: {
+      ...settings.attributes,
+      ...attributes,
+      ...( !! settings?.supports?.novaBlocks?.colorSignal?.altAttributes ? altAttributes : {} )
+    }
+  };
+}
+
+export default withColorSignalAttributes;
