+++ packages/block-editor/src/filters/with-color-signal/with-color-signal-deprecated.js
@@ -1,0 +1,35 @@
+const withColorSignalsDeprecated = ( settings, name ) => {
+
+  if ( ! settings?.supports?.novaBlocks?.colorSignal?.addOverlayColorDeprecatedMethod ) {
+    return settings;
+  }
+
+  return Object.assign( {}, settings, {
+    attributes: _.omit( settings.attributes, [ 'contentColor', 'overlayFilterStyle' ] ),
+    deprecated: [
+      {
+        attributes: settings.attributes,
+        isEligible( attributes ) {
+          return "undefined" === typeof attributes.paletteVariation;
+        },
+        migrate( attributes ) {
+          const { contentColor, overlayFilterStyle } = attributes;
+          const hadDarkText = !! contentColor && contentColor.search( '000' ) > -1;
+          const hadLightBackground = !! overlayFilterStyle && overlayFilterStyle === 'light';
+          const paletteVariation = hadDarkText || hadLightBackground ? '0' : '12';
+          const colorSignal = hadDarkText || hadLightBackground ? 0 : 3;
+
+          return {
+            ...attributes,
+            defaultsGenerated: true,
+            paletteVariation,
+            colorSignal,
+          };
+        },
+        save: settings.save,
+      }
+    ].concat( settings.deprecated ),
+  } );
+}
+
+export default withColorSignalsDeprecated;
