+++ packages/block-editor/src/filters/with-color-signal/with-color-signal-edit-classnames.js
@@ -1,0 +1,27 @@
+import classnames from "classnames";
+import { createHigherOrderComponent } from "@wordpress/compose";
+import { getColorSignalClassnames } from "@novablocks/utils";
+import { useSupports } from "../../hooks";
+
+const withColorSignalEditClassnames = createHigherOrderComponent( ( BlockListBlock ) => {
+
+  return ( props ) => {
+    const supports = useSupports( props.name );
+    const { attributes } = props;
+    const newClassnames = getColorSignalClassnames( attributes, supports );
+
+    const newProps = {
+      ...props,
+      className: classnames(
+        props.className,
+        newClassnames
+      )
+    };
+
+    return (
+      <BlockListBlock { ...newProps } />
+    )
+  };
+}, "withColorSignalEditClassnames" );
+
+export default withColorSignalEditClassnames;
