+++ packages/core/src/blocks/core/group/index.js
@@ -35,7 +35,7 @@ const alterSettings = ( settings ) => {
       novaBlocks: {
         colorSignal: {
           paletteClassname: true,
-          variationClassname: true,
+          paletteVariationClassname: true,
           colorSignalClassname: true,
           functionalColors: true,
         },
